refactor(agents): share parseAgentOutput helper across adapters
The parseAgentJson + validateAgentOutput fallback shape was duplicated
identically in parsePiOutput, parseOpenCodeOutput, and parseCopilotOutput
(only the function name and the no-JSON error string differed). Extract a
single parseAgentOutput(text, schema, agentLabel) next to
validateAgentOutput in types.ts and route all three adapters through it.

Behavior is preserved: the per-agent SyntaxError message is retained via
agentLabel, and each adapter keeps its own caller-side error formatting
(pi distinguishes Failed vs Invalid; opencode/copilot use a single message).
